requirements/base.txt
amqp==5.3.1
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
billiard==4.2.1
celery==5.3.6
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.39.1
cron_descriptor==2.0.5
cryptography==45.0.6
defusedxml==0.7.1
Django==4.2.11
django-appconf==1.1.0
django-autoslug==1.9.9
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-countries==7.5.1
django-filter==24.1
django-phonenumber-field==7.3.0
django-taggit==5.0.1
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.1
djoser==2.2.2
drf-yasg==1.21.7
flower==2.0.1
humanize==4.13.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
phonenumbers==8.13.32
pillow==10.2.0
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
